requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
